RavenDB-18675 Queries
