+++ src/Corax/Queries/AllEntriesMatch.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Reflection;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Voron;
 using Voron.Data.Containers;
 using Voron.Data.Sets;
-using Voron.Global;
 using Voron.Impl;
 
 namespace Corax.Queries
