+++ src/Corax/Queries/IQueryMatch.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Diagnostics;
 
 namespace Corax.Queries
 {
