+++ src/Corax/Queries/MultiTermMatch/TermProviders/TermProvider.EndsWith.cs
@@ -1,28 +1,26 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Sparrow.Server;
 using Voron;
 using Voron.Data.CompactTrees;
 
-namespace Corax.Queries
+namespace Corax.Queries.MultiTermMatch.TermProviders
 {
-    public unsafe struct NotContainsTermProvider : ITermProvider
+    public struct EndsWithTermProvider : ITermProvider
     {
         private readonly CompactTree _tree;
         private readonly IndexSearcher _searcher;
         private readonly Slice _fieldName;
-        private readonly Slice _term;
-
+        private readonly Slice _endsWith;
+        
         private CompactTree.Iterator _iterator;
-
-        public NotContainsTermProvider(IndexSearcher searcher, ByteStringContext context, CompactTree tree, Slice fieldName, int fieldId, Slice term)
+        public EndsWithTermProvider(IndexSearcher searcher, ByteStringContext context, CompactTree tree, Slice fieldName, int fieldId, Slice endsWith)
         {
             _tree = tree;
             _searcher = searcher;
             _fieldName = fieldName;
             _iterator = tree.Iterate();
             _iterator.Reset();
-            _term = term;
+            _endsWith = endsWith;
         }
 
         public void Reset()
@@ -33,10 +31,10 @@ namespace Corax.Queries
 
         public bool Next(out TermMatch term)
         {
-            var contains = _term;
+            var suffix = _endsWith;
             while (_iterator.MoveNext(out Slice termSlice, out var _))
             {
-                if (termSlice.Contains(contains))
+                if (termSlice.EndsWith(suffix) == false)
                     continue;
 
                 term = _searcher.TermQuery(_tree, termSlice);
@@ -49,12 +47,12 @@ namespace Corax.Queries
 
         public QueryInspectionNode Inspect()
         {
-            return new QueryInspectionNode($"{nameof(NotContainsTermProvider)}",
-                            parameters: new Dictionary<string, string>()
-                            {
-                                { "Field", _fieldName.ToString() },
-                                { "Term", _term.ToString()}
-                            });
+            return new QueryInspectionNode($"{nameof(EndsWithTermProvider)}",
+                parameters: new Dictionary<string, string>()
+                {
+                    { "Field", _fieldName.ToString() },
+                    { "Suffix", _endsWith.ToString()}
+                });
         }
     }
 }
