+++ src/Corax/Queries/MultiTermMatch/TermProviders/TermProvider.StartWith.cs
@@ -1,15 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Text;
-using System.Threading.Tasks;
 using Sparrow.Server;
 using Voron;
 using Voron.Data.CompactTrees;
 
-namespace Corax.Queries
+namespace Corax.Queries.MultiTermMatch.TermProviders
 {
     [DebuggerDisplay("{DebugView,nq}")]
     public struct StartWithTermProvider : ITermProvider
